correction zoom navigateur

+++ Module08_TempsInteractif/DSED_M08_DemoCours/DSED_M08_DemoCours/Entite/Point2d.cs
@@ -7,7 +7,7 @@ namespace DSED_M08_DemoCours.Entite
 {
     public class Point2d
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
     }
 }
